Add other weather data and their icons

+++ app/src/main/java/com/ugurbayrak/weatherapp/data/remote/dto/weather/Sys.kt
@@ -1,6 +1,6 @@
 package com.ugurbayrak.weatherapp.data.remote.dto.weather
 
 data class Sys(
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Long,
+    val sunset: Long
 )
